Compile fix for 2.0; see Mono bug #80384
2006-12-29  Alp Toker  <alp@atoker.com>

	* src/Banshee.Base/Banshee.Burner/BurnerSessionPreparer.cs:
	Compile fix for 2.0; see Mono bug #80384

+++ src/Banshee.Base/Banshee.Burner/BurnerSessionPreparer.cs
@@ -174,7 +174,7 @@ namespace Banshee.Burner
             if(free_space >= 0 && estimated_encoded_bytes >= free_space) {
                 LogCore.Instance.PushError(Catalog.GetString("Insufficient Disk Space"),
                     String.Format(Catalog.GetString("Creating this CD requires at least {0} MiB of free disk space."),
-                        Math.Ceiling(estimated_encoded_bytes / 1000000)));
+                        Math.Ceiling((double)estimated_encoded_bytes / 1000000)));
                 return false;
             }
